+++ maps/src/commonMain/kotlin/uz/yalla/maps/provider/google/GoogleStaticMap.kt
@@ -96,8 +96,6 @@ class GoogleStaticMap : StaticMap {
 
             if (!locations.isNullOrEmpty()) {
                 LocationsLayer(
-                    arrival = null,
-                    duration = null,
                     locations = locations,
                     startLabel = startLabel,
                     endLabel = endLabel